fix(v2): spool receiver host-guard + body cap
The OTLP spool receiver had no Host-header guard, so a DNS-rebinding /
browser page could POST telemetry into the spool, and an unbounded body
could exhaust disk / OOM the whole-file ingest read.

Extract isAllowedHost into a dep-free host-guard module shared with the
dashboard server; the spool fetch returns 403 on a present non-loopback
Host, and Bun.serve caps the body at OTLP_MAX_BODY_BYTES (8 MB).
